# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Second-Hand Store API",
    "description": "Fully autonomous second-hand store with AI negotiation",
    "version": "1.0.0"
  },
  "paths": {
    "/user/{user_id}/account": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Account Status",
        "description": "Lightweight probe used by the login flow. `verify_user_token` already raises\n403 for a banned account, so reaching the body means the user is allowed in.",
        "operationId": "account_status_user__user_id__account_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/user/{user_id}/password": {
      "put": {
        "tags": [
          "User"
        ],
        "summary": "Change Password",
        "description": "Change the authenticated user's password (verifies the current password first).",
        "operationId": "change_password_user__user_id__password_put",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordUpdateSchema"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/user/{user_id}/email": {
      "put": {
        "tags": [
          "User"
        ],
        "summary": "Change Email",
        "description": "Request a change of the authenticated user's email address.\n\nSPEC-056 #3. This used to hand `{\"email\": ..., \"email_confirm\": True}` to the\nadmin API with no password check: the new address was marked verified on the\nspot, and neither the old nor the new inbox heard about it. A stolen access\ntoken was therefore a permanent account takeover \u2014 the attacker owns the\naddress that password resets go to, and the real owner has no notification\nand no way back.\n\nNow it costs the password, and the update runs on the user's OWN session, so\nSupabase treats it as a change request and mails the confirmation link. The\naddress does not move until someone clicks it.",
        "operationId": "change_email_user__user_id__email_put",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailUpdateSchema"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/user/{user_id}/language": {
      "put": {
        "tags": [
          "User"
        ],
        "summary": "Update Language",
        "description": "Update the authenticated user's preferred language in Supabase auth user_metadata.\nSupported languages: 'en' (English), 'ms' (Bahasa Melayu), 'zh' (Simplified Chinese).",
        "operationId": "update_language_user__user_id__language_put",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LanguageUpdateSchema"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/user/{user_id}/profile": {
      "put": {
        "tags": [
          "User"
        ],
        "summary": "Update Profile",
        "description": "Update the authenticated user's display name and/or profile picture.\n\nThe avatar is uploaded to storage server-side using the service role key -\nclients are never allowed to write to the bucket directly (read-only).\nBoth values are stored in the Supabase auth user_metadata, the avatar under\n`custom_avatar_url` so an OAuth sign-in can't overwrite it.",
        "operationId": "update_profile_user__user_id__profile_put",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_update_profile_user__user_id__profile_put"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/user/{user_id}": {
      "delete": {
        "tags": [
          "User"
        ],
        "summary": "Delete Account",
        "description": "Permanently delete the authenticated user's account.\n\nRemoves app-side data (profile, chat settings, conversations) and the\nSupabase auth user. Orders/transactions are intentionally retained as\nbusiness records.\n\nSPEC-056 #7: costs the password too. This is irreversible and takes the\ntranscript with it, so a lifted access token must not be enough on its own.",
        "operationId": "delete_account_user__user_id__delete",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountDeleteSchema"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/payment/orders/user/{user_id}": {
      "get": {
        "tags": [
          "Payment"
        ],
        "summary": "Get User Orders",
        "description": "Get all orders for a specific user by their ID.\nReturns orders with item details.",
        "operationId": "get_user_orders_payment_orders_user__user_id__get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/users/{user_id}/profile": {
      "put": {
        "tags": [
          "System"
        ],
        "summary": "Update User Profile",
        "description": "Update a user's profile (display name, avatar).",
        "operationId": "update_user_profile_admin_users__user_id__profile_put",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserProfileUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/users/{user_id}": {
      "delete": {
        "tags": [
          "System"
        ],
        "summary": "Admin Delete User",
        "description": "Permanently delete a user: app-side data + the Supabase auth user.\nOrders are intentionally retained as business records.",
        "operationId": "admin_delete_user_admin_users__user_id__delete",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AccountDeleteSchema": {
        "properties": {
          "current_password": {
            "type": "string",
            "title": "Current Password"
          }
        },
        "type": "object",
        "required": [
          "current_password"
        ],
        "title": "AccountDeleteSchema"
      },
...
      "Body_update_profile_user__user_id__profile_put": {
        "properties": {
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "avatar": {
            "anyOf": [
              {
                "type": "string",
                "contentMediaType": "application/octet-stream"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar"
          }
        },
        "type": "object",
        "title": "Body_update_profile_user__user_id__profile_put"
      },
...
      "EmailUpdateSchema": {
        "properties": {
          "new_email": {
            "type": "string",
            "title": "New Email"
          },
          "current_password": {
            "type": "string",
            "title": "Current Password"
          }
        },
        "type": "object",
        "required": [
          "new_email",
          "current_password"
        ],
        "title": "EmailUpdateSchema"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LanguageUpdateSchema": {
        "properties": {
          "language": {
            "type": "string",
            "title": "Language"
          }
        },
        "type": "object",
        "required": [
          "language"
        ],
        "title": "LanguageUpdateSchema"
      },
...
      "PasswordUpdateSchema": {
        "properties": {
          "current_password": {
            "type": "string",
            "title": "Current Password"
          },
          "new_password": {
            "type": "string",
            "title": "New Password"
          }
        },
        "type": "object",
        "required": [
          "current_password",
          "new_password"
        ],
        "title": "PasswordUpdateSchema"
      },
...
      "UserProfileUpdateRequest": {
        "properties": {
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          }
        },
        "type": "object",
        "title": "UserProfileUpdateRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}